[#739] Prevent crashes due to question option without any actual options
